--- a/src/main/groovy/org/shipkit/internal/notes/contributors/AllContributorsFetcher.java
+++ b/src/main/groovy/org/shipkit/internal/notes/contributors/AllContributorsFetcher.java
@@ -45,6 +45,8 @@ class AllContributorsFetcher {
 
     private Set<ProjectContributor> extractContributors(List<JsonObject> page, final String readOnlyAuthToken) throws IOException, DeserializationException {
         Set<ProjectContributor> result = new HashSet<ProjectContributor>();
+        //Since returned contributor does not have 'name' element, we need to fetch the user data to get his name
+        //TODO add static caching of this. Names don't change that often, let's just cache this forever in build cache.
         ExecutorService executor = Executors.newFixedThreadPool(4);
         int maxSizePerCallable = 25;