--- a/src/main/groovy/org/shipkit/internal/notes/contributors/AllContributorsFetcher.java
+++ b/src/main/groovy/org/shipkit/internal/notes/contributors/AllContributorsFetcher.java
@@ -5,16 +5,11 @@ import org.gradle.api.logging.Logging;
 import org.json.simple.DeserializationException;
 import org.json.simple.JsonObject;
 import org.shipkit.internal.notes.model.ProjectContributor;
-import org.shipkit.internal.notes.util.Function;
 import org.shipkit.internal.notes.util.GitHubListFetcher;
-import org.shipkit.internal.notes.util.GitHubObjectFetcher;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.concurrent.*;
 
 /**
  * Gets all contributors from the repository
@@ -44,35 +39,9 @@ class AllContributorsFetcher {
     }
 
     private Set<ProjectContributor> extractContributors(List<JsonObject> page, final String readOnlyAuthToken) throws IOException, DeserializationException {
-        Set<ProjectContributor> result = new HashSet<ProjectContributor>();
         //Since returned contributor does not have 'name' element, we need to fetch the user data to get his name
         //TODO add static caching of this. Names don't change that often, let's just cache this forever in build cache.
-        ExecutorService executor = Executors.newFixedThreadPool(4);
-        int maxSizePerCallable = 25;
-
-        List<Future<Set<ProjectContributor>>> futures = new ArrayList<Future<Set<ProjectContributor>>>();
-        if (page.size() > 0) {
-            for (int i = 0; i < Math.max(page.size() / maxSizePerCallable, 1); i++) {
-                List<JsonObject> subList = page.subList(i * maxSizePerCallable, Math.min((i + 1) * maxSizePerCallable, page.size()));
-
-                GitHubObjectFetcher objectFetcher = new GitHubObjectFetcher(readOnlyAuthToken);
-                Function<JsonObject, ProjectContributor> projectContributorFetcherFunction = new ProjectContributorFetcherFunction(objectFetcher);
-                Callable<Set<ProjectContributor>> callable = new FetcherCallable<JsonObject, ProjectContributor>(subList, projectContributorFetcherFunction);
-
-                futures.add(executor.submit(callable));
-            }
-        }
-
-        for (Future<Set<ProjectContributor>> future: futures) {
-            try {
-                result.addAll(future.get());
-            } catch (InterruptedException e) {
-                throw new IOException(e);
-            } catch (ExecutionException e) {
-                throw new IOException(e.getCause());
-            }
-        }
-
+        Set<ProjectContributor> result = new ContributorsDispatcher().dispatch(page, readOnlyAuthToken);
         return result;
     }