--- a/src/main/groovy/org/shipkit/internal/notes/contributors/AllContributorsFetcher.java
+++ b/src/main/groovy/org/shipkit/internal/notes/contributors/AllContributorsFetcher.java
@@ -5,7 +5,9 @@ import org.gradle.api.logging.Logging;
 import org.json.simple.DeserializationException;
 import org.json.simple.JsonObject;
 import org.shipkit.internal.notes.model.ProjectContributor;
+import org.shipkit.internal.notes.util.Function;
 import org.shipkit.internal.notes.util.GitHubListFetcher;
+import org.shipkit.internal.notes.util.GitHubObjectFetcher;
 
 import java.io.IOException;
 import java.util.List;
@@ -41,7 +43,10 @@ class AllContributorsFetcher {
     private Set<ProjectContributor> extractContributors(List<JsonObject> page, final String readOnlyAuthToken) throws IOException, DeserializationException {
         //Since returned contributor does not have 'name' element, we need to fetch the user data to get his name
         //TODO add static caching of this. Names don't change that often, let's just cache this forever in build cache.
-        Set<ProjectContributor> result = new ContributorsDispatcher().dispatch(page, readOnlyAuthToken);
+        GitHubObjectFetcher objectFetcher = new GitHubObjectFetcher(readOnlyAuthToken);
+        Function<JsonObject, ProjectContributor> projectContributorFetcherFunction = new ProjectContributorFetcherFunction(objectFetcher);
+
+        Set<ProjectContributor> result = new ConcurrentDispatcher().dispatch(projectContributorFetcherFunction, page);
         return result;
     }