--- a/src/main/groovy/org/shipkit/internal/notes/util/Function.java
+++ b/src/main/groovy/org/shipkit/internal/notes/util/Function.java
@@ -1,0 +1,6 @@
+package org.shipkit.internal.notes.util;
+
+public interface Function<T, R> {
+
+    R apply(T t);
+}