--- a/src/test/groovy/org/shipkit/internal/notes/contributors/ProjectContributorFetcherFunctionTest.groovy
+++ b/src/test/groovy/org/shipkit/internal/notes/contributors/ProjectContributorFetcherFunctionTest.groovy
@@ -22,5 +22,7 @@ class ProjectContributorFetcherFunctionTest extends Specification {
         then:
         contributor
         contributor.login == 'epeee'
+        contributor.name
+        contributor.name.contains('Erhard')
     }
 }