for real world projects like mockito.

Currently, `fetchAllContributors` task takes ~22s while running `gradlew build` in mockito project.

The changes in this PR will end up in a duration of ~7s for the same task executed on the same project.

Note: this is still a very first version which might need some improvements (e.g the splitting up into sublists can be done smarter and also the error handling should be enhanced if we decide to merge the pr).

Actual profiling results:

 - using the currently released version:
`:fetchAllContributors 	22.492s` 	
 - using the version from this PR:
`:fetchAllContributors 	7.665s` 	

Furthermore, I did a diff on the resulting json files and it turned out that they are identical.